Agregado función para conectar a memcached sólo si se necesita.
git-svn-id: file:///home/gallir/meneame/svn/src@906 cbe1a1ec-7179-4919-8e32-4551e8279a00

+++ branches/version2/www/libs/utils.php
@@ -541,36 +541,41 @@ function fork($uri) {
 // Memcache functions
 //
 
-if ($globals['memcache_host']) {
-	$memcache = new Memcache;
-	if ( ! @$memcache->pconnect($globals['memcache_host'], 11211) ) {
-		$memcache = false;
-		syslog(LOG_INFO, "Meneame, memcache init failed");
+$memcache = false;
+
+function memcache_minit () {
+	global $memcache, $globals;
+
+	if ($memcache) return true;
+	if ($globals['memcache_host']) {
+		$memcache = new Memcache;
+		if ( ! @$memcache->pconnect($globals['memcache_host'], 11211) ) {
+			$memcache = false;
+			syslog(LOG_INFO, "Meneame, memcache init failed");
+			return false;
+		}
+		return true;
 	}
-} else {
-	$memcache = false;
+	return false;
 }
 
 function memcache_mget ($key) {
 	global $memcache;
-	if ($memcache) {
-		return $memcache->get($key);
-	}
+
+	if (memcache_minit()) return $memcache->get($key);
 	return false;
 }
 
 
 function memcache_madd ($key, $str, $expire=0) {
 	global $memcache;
-	if ($memcache) {
-		return $memcache->add($key, $str, false, $expire);
-	}
+	if (memcache_minit()) return $memcache->add($key, $str, false, $expire);
 	return false;
 }
 
 function memcache_mprint ($key) {
 	global $memcache;
-	if ($memcache && ($value = $memcache->get($key))) {
+	if (memcache_minit() && ($value = $memcache->get($key))) {
 		echo $value;
 		return true;
 	}
@@ -579,9 +584,7 @@ function memcache_mprint ($key) {
 
 function memcache_mdelete ($key) {
 	global $memcache;
-	if ($memcache) {
-		return $memcache->delete($key);
-	}
+	if (memcache_minit()) return $memcache->delete($key);
 	return false;
 }
 ?>
